Migrate GamblerParams to Axelrod
