+++ bin/pso_evolve.py
@@ -7,16 +7,20 @@ Based on Martin Jones @mojones original LookerUp code
 
 Usage:
     pso_evolve.py [-h] [--generations GENERATIONS] [--population POPULATION]
-    [--output OUTPUT_FILE] [--objective OBJECTIVE]
-    [--repetitions REPETITIONS] [--turns TURNS] [--noise NOISE]
-    [--nmoran NMORAN]
+    [--mu MUTATION_RATE] [--bottleneck BOTTLENECK] [--processes PROCESSES]
+    [--output OUTPUT_FILE] [--objective OBJECTIVE] [--repetitions REPETITIONS]
+    [--turns TURNS] [--noise NOISE] [--nmoran NMORAN]
     [--plays PLAYS] [--op_plays OP_PLAYS] [--op_start_plays OP_START_PLAYS]
+    [--algorithm ALGORITHM]
 
 Options:
     -h --help                   Show help
     --generations GENERATIONS   Generations to run the EA [default: 100]
     --population POPULATION     Starting population size  [default: 40]
-    --output OUTPUT_FILE        File to write data to [default: pso_tables.csv]
+    --mu MUTATION_RATE          Mutation rate [default: 0.1]
+    --bottleneck BOTTLENECK     Number of individuals to keep from each generation [default: 10]
+    --processes PROCESSES       Number of processes to use [default: 1]
+    --output OUTPUT_FILE        File to write data to [default: lookup_params.csv]
     --objective OBJECTIVE       Objective function [default: score]
     --repetitions REPETITIONS   Repetitions in objective [default: 100]
     --turns TURNS               Turns in each match [default: 200]
@@ -25,45 +29,34 @@ Options:
     --plays PLAYS               Number of recent plays in the lookup table [default: 2]
     --op_plays OP_PLAYS         Number of recent plays in the lookup table [default: 2]
     --op_start_plays OP_START_PLAYS     Number of opponent starting plays in the lookup table [default: 2]
+    --algorithm ALGORITHM       Which algorithm to use (EA for evolutionary algorithm or PS for
+                                particle swarm algorithm) [default: PS]
 """
 
-from docopt import docopt
 
-from axelrod_dojo.archetypes.gambler import GamblerParams
-from axelrod_dojo.algorithms.particle_swarm_optimization import PSO
+from axelrod import Action, Gambler
+from axelrod_dojo import invoke
 
 
-from axelrod_dojo.utils import prepare_objective
-
-if __name__ == "__main__":
-    arguments = docopt(__doc__, version='PSO Evolve 0.3')
-    print(arguments)
-
-    # Population Args
-    population = int(arguments['--population'])
-    generations = int(arguments['--generations'])
-
-    # Objective
-    name = str(arguments['--objective'])
-    repetitions = int(arguments['--repetitions'])
-    turns = int(arguments['--turns'])
-    noise = float(arguments['--noise'])
-    nmoran = int(arguments['--nmoran'])
-
-    # Lookup Tables
+def prepare_player_class_kwargs(arguments):
     plays = int(arguments['--plays'])
     op_plays = int(arguments['--op_plays'])
     op_start_plays = int(arguments['--op_start_plays'])
-    params_kwargs = {"plays": plays,
-                     "op_plays": op_plays,
-                     "op_state_plays": op_start_plays}
+    table_depth = max(plays, op_plays, op_start_plays)
+    initial_actions = [Action.C] * table_depth
+    param_kwargs = {
+        "parameters": (plays, op_plays, op_start_plays),
+        "initial_actions": initial_actions,
+        "mutation_probability": float(arguments['--mu']),
+    }
+    return param_kwargs
 
-    objective = prepare_objective(name, turns, noise, repetitions, nmoran)
 
-    pso = PSO(GamblerParams, params_kwargs, objective=objective,
-              population=population, generations=generations)
+if __name__ == '__main__':
+    invoke(
+        __doc__,
+        'PSO Evolver 0.4',
+        Gambler,
+        prepare_player_class_kwargs
+    )
 
-    xopt, fopt = pso.swarm()
-
-    print(xopt)
-    print(fopt)
