+++ src/axelrod_dojo/__init__.py
@@ -1,10 +1,7 @@
 from .version import __version__
 from .arguments import invoke
-from .archetypes.gambler import GamblerParams
 from .archetypes.cycler import CyclerParams
 from .algorithms.genetic_algorithm import Population
 from .algorithms.particle_swarm_optimization import PSO
-from .utils import (prepare_objective,
-                    load_params,
-                    PlayerInfo)
+from .utils import prepare_objective, load_params, PlayerInfo
